Change `engine_boots` and `engine_time` type to `u32`
It is possible to try to assign invalid values to
`SecurityParams.engine_boots` and `SecurityParams.engine_time` and the
setters will silently change the values. This could lead to hard to
track bugs.

Transfer the responsibility to check for invalid values to the code
using the crate. Add comments indicating what is the expected range.
This will give the opportunity to users of the crate to tolerate invalid
values.